chore: rename float -> decimal

+++ core/src/parser/node.rs
@@ -95,8 +95,8 @@ mod test {
             panic!("expected UnaryOp, got {:#?}; self is {:#?}", other.token, self.token);
           }
         }
-        Token::Number(N::Float(f1)) => {
-          if let Some(N::Float(f2)) = other.token.as_number() {
+        Token::Number(N::Decimal(f1)) => {
+          if let Some(N::Decimal(f2)) = other.token.as_number() {
             assert!((f1 - f2).abs() < Decimal::MAX);
           } else {
             panic!("expected N, got {:#?}; self is {:#?}", other.token, self.token)
